mcp_server/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
fastmcp
pydantic
mcp-cli